Update CreateFolderStructure.bat
This script uses the for /f "usebackq delims=" to read the lines in the file, it reads each line in the text file, and for /f "tokens=1-* delims= " to iterate through each token in the line. The script then creates the folder in the current folder's path, if it does not exist. The script uses the setlocal enabledelayedexpansion which is used to enable the use of !var! instead of %var% in the for loops in order to expand variables.